requirements.txt
flask
mlflow
torch
torchvision
Pillow
